end12

+++ controller/Products.js
@@ -208,7 +208,12 @@ export const create = async (req, res) => {
 
 export const deletes = async (req, res) => {
 
-    
+   const { email } = req.user;
+   const user = await User.findOne({ email });
+
+   if (user.isAdmin === false) {
+      return res.status(400).json({ success: false, msg: 'Only admin can delete product' });
+   }
    try {
       const id = req.params.id;
       const product = await Product.findByIdAndDelete(id);
@@ -223,12 +228,12 @@ export const deletes = async (req, res) => {
 
 export const update = async (req, res) => {
 
-      const {email} = req.user;
-      const user = await User.findOne({email});
-   
-      if(user.isAdmin === false){
-         return res.status(400).json({success : false,msg : 'Only admin can edit product'});
-      }
+   const { email } = req.user;
+   const user = await User.findOne({ email });
+
+   if (user.isAdmin === false) {
+      return res.status(400).json({ success: false, msg: 'Only admin can edit product' });
+   }
    try {
       const isvalidcategory = await Categories.findById(req.body.category);
       if (!isvalidcategory) {
